[NOID] updated license files
