+++ full-it/src/test/java/apoc/full/it/vectordb/ChromaDbTest.java
@@ -1,4 +1,4 @@
-package apoc.vectordb;
+package apoc.full.it.vectordb;
 
 import apoc.util.TestUtil;
 import apoc.util.Util;
@@ -12,52 +12,42 @@ import org.junit.rules.TemporaryFolder;
 import org.neo4j.dbms.api.DatabaseManagementService;
 import org.neo4j.graphdb.GraphDatabaseService;
 import org.neo4j.test.TestDatabaseManagementServiceBuilder;
-import org.testcontainers.qdrant.QdrantContainer;
+import org.testcontainers.chromadb.ChromaDBContainer;
 
+import java.util.List;
 import java.util.Map;
+import java.util.concurrent.atomic.AtomicReference;
 
-import static apoc.ml.RestAPIConfig.HEADERS_KEY;
 import static apoc.util.MapUtil.map;
 import static apoc.util.TestUtil.testCall;
 import static apoc.util.TestUtil.testResult;
-import static apoc.vectordb.VectorDbHandler.Type.QDRANT;
-import static apoc.vectordb.VectorDbTestUtil.EntityType.NODE;
-import static apoc.vectordb.VectorDbTestUtil.EntityType.FALSE;
-import static apoc.vectordb.VectorDbTestUtil.EntityType.REL;
+import static apoc.vectordb.VectorDbHandler.Type.CHROMA;
 import static apoc.vectordb.VectorDbTestUtil.assertBerlinResult;
 import static apoc.vectordb.VectorDbTestUtil.assertLondonResult;
 import static apoc.vectordb.VectorDbTestUtil.assertNodesCreated;
 import static apoc.vectordb.VectorDbTestUtil.assertRelsCreated;
 import static apoc.vectordb.VectorDbTestUtil.dropAndDeleteAll;
-import static apoc.vectordb.VectorDbTestUtil.getAuthHeader;
+import static apoc.vectordb.VectorDbTestUtil.EntityType.*;
 import static apoc.vectordb.VectorDbUtil.ERROR_READONLY_MAPPING;
 import static apoc.vectordb.VectorEmbeddingConfig.ALL_RESULTS_KEY;
 import static apoc.vectordb.VectorEmbeddingConfig.MAPPING_KEY;
 import static apoc.vectordb.VectorMappingConfig.*;
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.fail;
-import static org.neo4j.configuration.GraphDatabaseSettings.*;
+import static org.neo4j.configuration.GraphDatabaseSettings.DEFAULT_DATABASE_NAME;
+import static org.neo4j.configuration.GraphDatabaseSettings.SYSTEM_DATABASE_NAME;
 
-public class QdrantTest {
-    private static final String ADMIN_KEY = "my_admin_api_key";
-    private static final String READONLY_KEY = "my_readonly_api_key";
-    
-    private static final QdrantContainer QDRANT_CONTAINER = new QdrantContainer("qdrant/qdrant:v1.7.4")
-            .withEnv("QDRANT__SERVICE__API_KEY", ADMIN_KEY)
-            .withEnv("QDRANT__SERVICE__READ_ONLY_API_KEY", READONLY_KEY);
-
-    private static final Map<String, String> ADMIN_AUTHORIZATION = getAuthHeader(ADMIN_KEY);
-    private static final Map<String, String> READONLY_AUTHORIZATION = getAuthHeader(READONLY_KEY);
-    private static final Map<String, Object> ADMIN_HEADER_CONF  = map(HEADERS_KEY, ADMIN_AUTHORIZATION);
+public class ChromaDbTest {
+    private static final AtomicReference<String> COLL_ID = new AtomicReference<>();
+    private static final ChromaDBContainer CHROMA_CONTAINER = new ChromaDBContainer("chromadb/chroma:0.4.25.dev137");
 
     private static String HOST;
 
     @ClassRule
     public static TemporaryFolder storeDir = new TemporaryFolder();
-    
+
     private static GraphDatabaseService sysDb;
     private static GraphDatabaseService db;
     private static DatabaseManagementService databaseManagementService;
@@ -69,57 +59,97 @@ public class QdrantTest {
         db = databaseManagementService.database(DEFAULT_DATABASE_NAME);
         sysDb = databaseManagementService.database(SYSTEM_DATABASE_NAME);
         
-        QDRANT_CONTAINER.start();
+        CHROMA_CONTAINER.start();
 
-        HOST = "localhost:" + QDRANT_CONTAINER.getMappedPort(6333);
-        TestUtil.registerProcedure(db, Qdrant.class, VectorDb.class);
-
-        testCall(db, "CALL apoc.vectordb.qdrant.createCollection($host, 'test_collection', 'Cosine', 4, $conf)",
-                map("host", HOST, "conf", ADMIN_HEADER_CONF),
+        HOST = "localhost:" + CHROMA_CONTAINER.getMappedPort(8000);
+        TestUtil.registerProcedure(db, ChromaDb.class, VectorDb.class);
+        
+        testCall(db, "CALL apoc.vectordb.chroma.createCollection($host, 'test_collection', 'cosine', 4)",
+            map("host", HOST),
                 r -> {
                     Map value = (Map) r.get("value");
-                    assertEquals("ok", value.get("status"));
+                    COLL_ID.set((String) value.get("id"));
                 });
 
         testCall(db, """
-                        CALL apoc.vectordb.qdrant.upsert($host, 'test_collection',
+                        CALL apoc.vectordb.chroma.upsert($host, $collection,
                         [
-                            {id: 1, vector: [0.05, 0.61, 0.76, 0.74], metadata: {city: "Berlin", foo: "one"}},
-                            {id: 2, vector: [0.19, 0.81, 0.75, 0.11], metadata: {city: "London", foo: "two"}}
-                        ],
-                        $conf)
+                            {id: '1', vector: [0.05, 0.61, 0.76, 0.74], metadata: {city: "Berlin", foo: "one"}, text: 'ajeje'},
+                            {id: '2', vector: [0.19, 0.81, 0.75, 0.11], metadata: {city: "London", foo: "two"}, text: 'brazorf'}
+                        ])
                         """,
-                map("host", HOST, "conf", ADMIN_HEADER_CONF),
+                map("host", HOST, "collection", COLL_ID.get()),
                 r -> {
-                    Map value = (Map) r.get("value");
-                    assertEquals("ok", value.get("status"));
+                    assertNull(r.get("value"));
                 });
     }
 
     @AfterClass
     public static void tearDown() throws Exception {
-        testCall(db, "CALL apoc.vectordb.qdrant.deleteCollection($host, 'test_collection', $conf)",
-                map("host", HOST, "conf", ADMIN_HEADER_CONF),
+        testCall(db, "CALL apoc.vectordb.chroma.deleteCollection($host, 'test_collection')",
+                map("host", HOST),
                 r -> {
                     Map value = (Map) r.get("value");
-                    assertEquals(true, value.get("result"));
+                    assertNull(value);
                 });
 
         databaseManagementService.shutdown();
-        QDRANT_CONTAINER.stop();
+        CHROMA_CONTAINER.stop();
     }
-    
+
     @Before
     public void before() {
         dropAndDeleteAll(db);
     }
     
     @Test
-    public void getVectorsWithReadOnlyApiKey() {
-        testResult(db, "CALL apoc.vectordb.qdrant.get($host, 'test_collection', [1], $conf) ",
-                map("host", HOST, 
-                        "conf", map(ALL_RESULTS_KEY, true, HEADERS_KEY, READONLY_AUTHORIZATION)
-                ),
+    public void getVectors() {
+        testResult(db, "CALL apoc.vectordb.chroma.get($host, $collection, ['1'], $conf) ",
+                map("host", HOST, "collection", COLL_ID.get(), "conf", map(ALL_RESULTS_KEY, true)),
+                r -> {
+                    Map<String, Object> row = r.next();
+                    assertBerlinResult(row, FALSE);
+                    assertNotNull(row.get("vector"));
+                    assertEquals("ajeje", row.get("text"));
+                });
+    }
+
+    @Test
+    public void getVectorsWithoutVectorResult() {
+        testResult(db, "CALL apoc.vectordb.chroma.get($host, $collection, ['1'])",
+                map("host", HOST, "collection", COLL_ID.get()),
+                r -> {
+                    Map<String, Object> row = r.next();
+                    assertEquals(Map.of("city", "Berlin", "foo", "one"), row.get("metadata"));
+                    assertNull(row.get("vector"));
+                    assertNull(row.get("id"));
+                });
+    }
+    
+    @Test
+    public void deleteVector() {
+        testCall(db, """
+                        CALL apoc.vectordb.chroma.upsert($host, $collection,
+                        [
+                            {id: 3, embedding: [0.19, 0.81, 0.75, 0.11], metadata: {foo: "baz"}}
+                        ])
+                        """,
+                map("host", HOST, "collection", COLL_ID.get()),
+                r -> {
+                    assertNull(r.get("value"));
+                });
+
+        testCall(db, "CALL apoc.vectordb.chroma.delete($host, $collection, [3]) ",
+                map("host", HOST, "collection", COLL_ID.get()),
+                r -> {
+                    assertEquals(List.of("3"), r.get("value"));
+                });
+    }
+
+    @Test
+    public void createAndDeleteVector() {
+        testResult(db, "CALL apoc.vectordb.chroma.get($host, $collection, ['1'], $conf) ",
+                map("host", HOST, "collection", COLL_ID.get(), "conf", map(ALL_RESULTS_KEY, true)),
                 r -> {
                     Map<String, Object> row = r.next();
                     assertBerlinResult(row, FALSE);
@@ -128,59 +158,9 @@ public class QdrantTest {
     }
 
     @Test
-    public void writeOperationWithReadOnlyUser() {
-        try {
-            testCall(db, "CALL apoc.vectordb.qdrant.deleteCollection($host, 'test_collection', $conf)",
-                    Util.map("host", HOST,
-                            "conf", Util.map(HEADERS_KEY, READONLY_AUTHORIZATION)
-                    ),
-                    r -> fail()
-            );
-        } catch (Exception e) {
-            assertThat( e.getMessage() ).contains("HTTP response code: 403");
-        }
-    }
-    
-    @Test
-    public void getVectorsWithoutVectorResult() {
-        testResult(db, "CALL apoc.vectordb.qdrant.get($host, 'test_collection', [1], $conf) ",
-                map("host", HOST, "conf", ADMIN_HEADER_CONF),
-                r -> {
-                    Map<String, Object> row = r.next();
-                    assertEquals(Map.of("city", "Berlin", "foo", "one"), row.get("metadata"));
-                    assertNull(row.get("vector"));
-                    assertNull(row.get("id"));
-                });
-    }
-
-    @Test
-    public void deleteVector() {
-        testCall(db, """
-                        CALL apoc.vectordb.qdrant.upsert($host, 'test_collection',
-                        [
-                            {id: 3, vector: [0.19, 0.81, 0.75, 0.11], metadata: {foo: "baz"}},
-                            {id: 4, vector: [0.19, 0.81, 0.75, 0.11], metadata: {foo: "baz"}}
-                        ],
-                        $conf)
-                        """,
-                map("host", HOST, "conf", ADMIN_HEADER_CONF),
-                r -> {
-                    Map value = (Map) r.get("value");
-                    assertEquals("ok", value.get("status"));
-                });
-        
-        testCall(db, "CALL apoc.vectordb.qdrant.delete($host, 'test_collection', [3, 4], $conf) ",
-                map("host", HOST, "conf", ADMIN_HEADER_CONF),
-                r -> {
-                    Map value = (Map) r.get("value");
-                    assertEquals("ok", value.get("status"));
-                });
-    }
-
-    @Test
     public void queryVectors() {
-        testResult(db, "CALL apoc.vectordb.qdrant.query($host, 'test_collection', [0.2, 0.1, 0.9, 0.7], {}, 5, $conf)",
-                map("host", HOST, "conf", map(ALL_RESULTS_KEY, true, HEADERS_KEY, ADMIN_AUTHORIZATION)),
+        testResult(db, "CALL apoc.vectordb.chroma.query($host, $collection, [0.2, 0.1, 0.9, 0.7], {}, 5, $conf)",
+                map("host", HOST, "collection", COLL_ID.get(), "conf", map(ALL_RESULTS_KEY, true)),
                 r -> {
                     Map<String, Object> row = r.next();
                     assertBerlinResult(row, FALSE);
@@ -193,11 +173,12 @@ public class QdrantTest {
                     assertNotNull(row.get("vector"));
                 });
     }
-
+    
     @Test
     public void queryVectorsWithoutVectorResult() {
-        testResult(db, "CALL apoc.vectordb.qdrant.query($host, 'test_collection', [0.2, 0.1, 0.9, 0.7], {}, 5, $conf)",
-                map("host", HOST, "conf", map(HEADERS_KEY, ADMIN_AUTHORIZATION)),
+        testResult(db, "CALL apoc.vectordb.chroma.query($host, $collection, [0.2, 0.1, 0.9, 0.7], {}, 5) " +
+                       " YIELD score, vector, id, metadata, node RETURN * ORDER BY id",
+                map("host", HOST, "collection", COLL_ID.get()),
                 r -> {
                     Map<String, Object> row = r.next();
                     assertEquals(Map.of("city", "Berlin", "foo", "one"), row.get("metadata"));
@@ -215,10 +196,8 @@ public class QdrantTest {
 
     @Test
     public void queryVectorsWithYield() {
-        testResult(db, "CALL apoc.vectordb.qdrant.query($host, 'test_collection', [0.2, 0.1, 0.9, 0.7], {}, 5, $conf) YIELD metadata, id",
-                map("host", HOST,
-                        "conf", map(ALL_RESULTS_KEY, true, HEADERS_KEY, ADMIN_AUTHORIZATION)
-                ),
+        testResult(db, "CALL apoc.vectordb.chroma.query($host, $collection, [0.2, 0.1, 0.9, 0.7], {}, 5, $conf) YIELD metadata, id",
+                map("host", HOST, "collection", COLL_ID.get(), "conf", map(ALL_RESULTS_KEY, true)),
                 r -> {
                     assertBerlinResult(r.next(), FALSE);
                     assertLondonResult(r.next(), FALSE);
@@ -228,14 +207,8 @@ public class QdrantTest {
     @Test
     public void queryVectorsWithFilter() {
         testResult(db, """
-                        CALL apoc.vectordb.qdrant.query($host, 'test_collection', [0.2, 0.1, 0.9, 0.7],
-                        { must:
-                            [ { key: "city", match: { value: "London" } } ]
-                        },
-                        5, $conf) YIELD metadata, id""",
-                map("host", HOST,
-                        "conf", map(ALL_RESULTS_KEY, true, HEADERS_KEY, ADMIN_AUTHORIZATION)
-                ),
+                        CALL apoc.vectordb.chroma.query($host, $collection, [0.2, 0.1, 0.9, 0.7], {city: 'London'}, 5, $conf) YIELD metadata, id""",
+                map("host", HOST, "collection", COLL_ID.get(), "conf", map(ALL_RESULTS_KEY, true)),
                 r -> {
                     assertLondonResult(r.next(), FALSE);
                 });
@@ -244,10 +217,8 @@ public class QdrantTest {
     @Test
     public void queryVectorsWithLimit() {
         testResult(db, """
-                        CALL apoc.vectordb.qdrant.query($host, 'test_collection', [0.2, 0.1, 0.9, 0.7], {}, 1, $conf) YIELD metadata, id""",
-                map("host", HOST,
-                        "conf", map(ALL_RESULTS_KEY, true, HEADERS_KEY, ADMIN_AUTHORIZATION)
-                ),
+                        CALL apoc.vectordb.chroma.query($host, $collection, [0.2, 0.1, 0.9, 0.7], {}, 1, $conf) YIELD metadata, id""",
+                map("host", HOST, "collection", COLL_ID.get(), "conf", map(ALL_RESULTS_KEY, true)),
                 r -> {
                     assertBerlinResult(r.next(), FALSE);
                 });
@@ -255,18 +226,16 @@ public class QdrantTest {
 
     @Test
     public void queryVectorsWithCreateNode() {
-
         Map<String, Object> conf = map(ALL_RESULTS_KEY, true,
-                HEADERS_KEY, ADMIN_AUTHORIZATION,
-                MAPPING_KEY, map(
-                        EMBEDDING_KEY, "vect",
-                        NODE_LABEL, "Test",
-                        ENTITY_KEY, "myId",
-                        METADATA_KEY, "foo",
-                        CREATE_KEY, true)
+                MAPPING_KEY, map(EMBEDDING_KEY, "vect",
+                    NODE_LABEL, "Test",
+                    ENTITY_KEY, "myId",
+                    METADATA_KEY, "foo",
+                    CREATE_KEY, true)
         );
-        testResult(db, "CALL apoc.vectordb.qdrant.queryAndUpdate($host, 'test_collection', [0.2, 0.1, 0.9, 0.7], {}, 5, $conf)",
-                map("host", HOST, "conf", conf),
+        
+        testResult(db, "CALL apoc.vectordb.chroma.queryAndUpdate($host, $collection, [0.2, 0.1, 0.9, 0.7], {}, 5, $conf)",
+                map("host", HOST, "collection", COLL_ID.get(), "conf", conf),
                 r -> {
                     Map<String, Object> row = r.next();
                     assertBerlinResult(row, NODE);
@@ -281,11 +250,9 @@ public class QdrantTest {
 
         assertNodesCreated(db);
 
-        testResult(db, "MATCH (n:Test) RETURN properties(n) AS props ORDER BY n.myId",
-                VectorDbTestUtil::vectorEntityAssertions);
-
-        testResult(db, "CALL apoc.vectordb.qdrant.queryAndUpdate($host, 'test_collection', [0.2, 0.1, 0.9, 0.7], {}, 5, $conf)",
-                map("host", HOST, "conf", conf),
+        testResult(db, "CALL apoc.vectordb.chroma.queryAndUpdate($host, $collection, [0.22, 0.11, 0.99, 0.17], {}, 5, $conf) " +
+                       "   YIELD score, vector, id, metadata, node RETURN * ORDER BY id",
+                map("host", HOST, "collection", COLL_ID.get(), "conf", conf),
                 r -> {
                     Map<String, Object> row = r.next();
                     assertBerlinResult(row, NODE);
@@ -307,15 +274,13 @@ public class QdrantTest {
         db.executeTransactionally("CREATE (:Test {myId: 'one'}), (:Test {myId: 'two'})");
 
         Map<String, Object> conf = map(ALL_RESULTS_KEY, true,
-                HEADERS_KEY, ADMIN_AUTHORIZATION,
                 MAPPING_KEY, map(EMBEDDING_KEY, "vect",
                         NODE_LABEL, "Test",
                         ENTITY_KEY, "myId",
                         METADATA_KEY, "foo"));
 
-        testResult(db, "CALL apoc.vectordb.qdrant.getAndUpdate($host, 'test_collection', [1, 2], $conf) " +
-                       "YIELD vector, id, metadata, node RETURN * ORDER BY id",
-                map("host", HOST, "conf", conf),
+        testResult(db, "CALL apoc.vectordb.chroma.getAndUpdate($host, $collection, ['1', '2'], $conf)",
+                map("host", HOST, "collection", COLL_ID.get(), "conf", conf),
                 r -> {
                     Map<String, Object> row = r.next();
                     assertBerlinResult(row, NODE);
@@ -335,29 +300,27 @@ public class QdrantTest {
                 MAPPING_KEY, map(EMBEDDING_KEY, "vect"));
 
         try {
-            testCall(db, "CALL apoc.vectordb.qdrant.get($host, 'test_collection', [1, 2], $conf)",
-                    map("host", HOST, "conf", conf),
+            testCall(db, "CALL apoc.vectordb.chroma.get($host, $collection, [1, 2], $conf)",
+                    map("host", HOST, "collection", COLL_ID.get(), "conf", conf),
                     r -> fail()
             );
         } catch (RuntimeException e) {
             Assertions.assertThat(e.getMessage()).contains(ERROR_READONLY_MAPPING);
         }
     }
-
+    
     @Test
     public void queryVectorsWithCreateNodeUsingExistingNode() {
 
         db.executeTransactionally("CREATE (:Test {myId: 'one'}), (:Test {myId: 'two'})");
 
         Map<String, Object> conf = map(ALL_RESULTS_KEY, true,
-                HEADERS_KEY, ADMIN_AUTHORIZATION,
                 MAPPING_KEY, map(EMBEDDING_KEY, "vect",
                         NODE_LABEL, "Test",
                         ENTITY_KEY, "myId",
                         METADATA_KEY, "foo"));
-        
-        testResult(db, "CALL apoc.vectordb.qdrant.queryAndUpdate($host, 'test_collection', [0.2, 0.1, 0.9, 0.7], {}, 5, $conf)",
-                map("host", HOST, "conf", conf),
+        testResult(db, "CALL apoc.vectordb.chroma.queryAndUpdate($host, $collection, [0.2, 0.1, 0.9, 0.7], {}, 5, $conf)",
+                map("host", HOST, "collection", COLL_ID.get(), "conf", conf),
                 r -> {
                     Map<String, Object> row = r.next();
                     assertBerlinResult(row, NODE);
@@ -374,20 +337,33 @@ public class QdrantTest {
     }
 
     @Test
+    public void queryReadOnlyVectorsWithMapping() {
+        Map<String, Object> conf = map(ALL_RESULTS_KEY, true,
+                MAPPING_KEY, map(EMBEDDING_KEY, "vect"));
+        
+        try {
+            testCall(db, "CALL apoc.vectordb.chroma.query($host, $collection, [0.2, 0.1, 0.9, 0.7], {}, 5, $conf)",
+                    map("host", HOST, "collection", COLL_ID.get(), "conf", conf),
+                    r -> fail()
+            );
+        } catch (RuntimeException e) {
+            Assertions.assertThat(e.getMessage()).contains(ERROR_READONLY_MAPPING);
+        }
+    }
+
+    @Test
     public void queryVectorsWithCreateRel() {
 
         db.executeTransactionally("CREATE (:Start)-[:TEST {myId: 'one'}]->(:End), (:Start)-[:TEST {myId: 'two'}]->(:End)");
-        
+
         Map<String, Object> conf = map(ALL_RESULTS_KEY, true,
-                HEADERS_KEY, ADMIN_AUTHORIZATION,
-                MAPPING_KEY, map(
-                        EMBEDDING_KEY, "vect",
-                        REL_TYPE, "TEST",
-                        ENTITY_KEY, "myId",
-                        METADATA_KEY, "foo")
-        );
-        testResult(db, "CALL apoc.vectordb.qdrant.queryAndUpdate($host, 'test_collection', [0.2, 0.1, 0.9, 0.7], {}, 5, $conf)",
-                map("host", HOST, "conf", conf),
+                MAPPING_KEY, map(EMBEDDING_KEY, "vect",
+                REL_TYPE, "TEST",
+                ENTITY_KEY, "myId",
+                METADATA_KEY, "foo",
+                "create", true));
+        testResult(db, "CALL apoc.vectordb.chroma.queryAndUpdate($host, $collection, [0.2, 0.1, 0.9, 0.7], {}, 5, $conf)",
+                map("host", HOST, "collection", COLL_ID.get(), "conf", conf),
                 r -> {
                     Map<String, Object> row = r.next();
                     assertBerlinResult(row, REL);
@@ -404,44 +380,29 @@ public class QdrantTest {
     }
 
     @Test
-    public void queryReadOnlyVectorsWithMapping() {
-        Map<String, Object> conf = map(ALL_RESULTS_KEY, true,
-                MAPPING_KEY, map(EMBEDDING_KEY, "vect"));
-
-        try {
-            testCall(db, "CALL apoc.vectordb.qdrant.query($host, 'test_collection', [0.2, 0.1, 0.9, 0.7], {}, 5, $conf)",
-                    map("host", HOST, "conf", conf),
-                    r -> fail()
-            );
-        } catch (RuntimeException e) {
-            Assertions.assertThat(e.getMessage()).contains(ERROR_READONLY_MAPPING);
-        }
-    }
-
-    @Test
     public void queryVectorsWithSystemDbStorage() {
-        String keyConfig = "qdrant-config-foo";
+        String keyConfig = "chroma-config-foo";
         String baseUrl = "http://" + HOST;
         Map<String, Object> mapping = map(EMBEDDING_KEY, "vect",
                 NODE_LABEL, "Test",
                 ENTITY_KEY, "myId",
                 METADATA_KEY, "foo");
-        sysDb.executeTransactionally("CALL apoc.vectordb.configure($vectorName, $keyConfig, $databaseName, $conf)", 
-                map("vectorName", QDRANT.toString(),
+        sysDb.executeTransactionally("CALL apoc.vectordb.configure($vectorName, $keyConfig, $databaseName, $conf)",
+                map("vectorName", CHROMA.toString(),
                         "keyConfig", keyConfig,
                         "databaseName", DEFAULT_DATABASE_NAME,
                         "conf", map(
-                            "host", baseUrl,
-                            "credentials", ADMIN_KEY,
-                            "mapping", mapping
+                                "host", baseUrl,
+                                "credentials", null,
+                                "mapping", mapping
                         )
                 )
         );
 
         db.executeTransactionally("CREATE (:Test {myId: 'one'}), (:Test {myId: 'two'})");
 
-        testResult(db, "CALL apoc.vectordb.qdrant.queryAndUpdate($host, 'test_collection', [0.2, 0.1, 0.9, 0.7], {}, 5, $conf)",
-                map("host", keyConfig, "conf", map(ALL_RESULTS_KEY, true)),
+        testResult(db, "CALL apoc.vectordb.chroma.queryAndUpdate($host, $collection, [0.2, 0.1, 0.9, 0.7], {}, 5, $conf)",
+                map("host", keyConfig, "collection", COLL_ID.get(), "conf", map(ALL_RESULTS_KEY, true)),
                 r -> {
                     Map<String, Object> row = r.next();
                     assertBerlinResult(row, NODE);
@@ -456,5 +417,4 @@ public class QdrantTest {
 
         assertNodesCreated(db);
     }
-
 }
